# support all type of slice in 'in' keyword