It will be slow, the benchmark is follow.

| benchmark        |  old ns/op           | new ns/op  | delta |
| ------------- |:-------------:| -----:|:---:|
| Benchmark_BuildIN-8      |                86.7       |   602      |     +594.35%|